requirements.txt
asgiref==3.8.1
attrs==25.1.0
colorama==0.4.6
Django==4.2
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
environ==1.0

geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1

python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
wheel==0.45.1
